Add concrete retry strategies
